+++ kork-web/src/test/groovy/com/netflix/spinnaker/kork/web/exceptions/GenericExceptionHandlersSpec.groovy
@@ -16,10 +16,6 @@
 
 package com.netflix.spinnaker.kork.web.exceptions
 
-import com.netflix.spinnaker.kork.api.exceptions.ExceptionDetails
-import com.netflix.spinnaker.kork.api.exceptions.ExceptionMessage
-import org.springframework.beans.BeansException
-import org.springframework.beans.factory.ObjectProvider
 import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.ResponseStatus
 import spock.lang.Shared
@@ -27,7 +23,6 @@ import spock.lang.Specification
 import spock.lang.Subject
 import spock.lang.Unroll
 
-import javax.annotation.Nullable
 import javax.servlet.http.HttpServletRequest
 import javax.servlet.http.HttpServletResponse;
 
@@ -65,7 +60,7 @@ class GenericExceptionHandlersSpec extends Specification {
     new E4()                                 || 400                || "My Other Reason!" // favor @ResponseStatus on interface over super class
     new E5("E5 Reason")                      || 400                || "E5 Reason"
     new NullPointerException("It's an NPE!") || 500                || "It's an NPE!"
-    new AccessDenied()                       || 403                || "Access is denied" + "\n" + messageToBeAppended
+    new LocalAccessDeniedException()         || 403                || "Access is denied" + "\n" + messageToBeAppended
   }
 
   @ResponseStatus(value = HttpStatus.NOT_FOUND, reason = "Default Reason!")
@@ -123,64 +118,6 @@ class GenericExceptionHandlersSpec extends Specification {
       super(message)
     }
   }
-
-  @ResponseStatus(value = HttpStatus.FORBIDDEN, reason = "Access is denied")
-  class AccessDenied extends E1 {
-    AccessDenied() {
-      super()
-    }
-
-    AccessDenied(String message) {
-      super(message)
-    }
-  }
-
 }
 
-class ExceptionMessageProvider implements ObjectProvider<List<ExceptionMessage>> {
 
-  List<ExceptionMessage> exceptionMessages
-
-  ExceptionMessageProvider(List<ExceptionMessage> exceptionMessages) {
-    this.exceptionMessages = exceptionMessages
-  }
-
-  @Override
-  List<ExceptionMessage> getObject(Object... args) throws BeansException {
-    return exceptionMessages
-  }
-
-  @Override
-  List<ExceptionMessage> getIfAvailable() throws BeansException {
-    return exceptionMessages
-  }
-
-  @Override
-  List<ExceptionMessage> getIfUnique() throws BeansException {
-    return exceptionMessages
-  }
-
-  @Override
-  List<ExceptionMessage> getObject() throws BeansException {
-    return exceptionMessages
-  }
-}
-
-class AccessDeniedExceptionMessage implements ExceptionMessage {
-
-  private String messageToBeAppended
-
-  AccessDeniedExceptionMessage(String messageToBeAppended) {
-    this.messageToBeAppended = messageToBeAppended
-  }
-
-  @Override
-  boolean supports(Class<? extends Throwable> throwable) {
-    return throwable == GenericExceptionHandlersSpec.AccessDenied.class
-  }
-
-  @Override
-  String message(Throwable throwable, @Nullable ExceptionDetails exceptionDetails) {
-    return messageToBeAppended
-  }
-}
