+++ kork-web/src/test/groovy/com/netflix/spinnaker/kork/web/exceptions/utils.groovy
@@ -22,39 +22,10 @@ import com.netflix.spinnaker.kork.api.exceptions.ExceptionMessage
 import com.netflix.spinnaker.kork.exceptions.SpinnakerException
 import org.springframework.beans.BeansException
 import org.springframework.beans.factory.ObjectProvider
-import spock.lang.Specification
+import org.springframework.http.HttpStatus
+import org.springframework.web.bind.annotation.ResponseStatus
 
 import javax.annotation.Nullable
-
-class ExceptionMessageDecoratorSpec extends Specification {
-
-  String messageToBeAppended = "Message to be appended."
-  AccessDeniedExceptionMessage accessDeniedExceptionMessage = new AccessDeniedExceptionMessage(messageToBeAppended)
-  ExceptionMessageProvider exceptionMessageProvider = new ExceptionMessageProvider([accessDeniedExceptionMessage])
-  ExceptionMessageDecorator exceptionMessageDecorator = new ExceptionMessageDecorator(exceptionMessageProvider)
-
-  def "Returns a message that is the original exception message and the message provided from the appender"() {
-    given:
-    LocalAccessDeniedException accessDeniedException = new LocalAccessDeniedException("Access is denied.")
-
-    when:
-    String userMessage = exceptionMessageDecorator.decorate(accessDeniedException, accessDeniedException.getMessage())
-
-    then:
-    userMessage == accessDeniedException.getMessage() + "\n" + messageToBeAppended
-  }
-
-  def "Does not return an appended message when the exception type is unsupported"() {
-    given:
-    RuntimeException runtimeException = new RuntimeException("Runtime exception.")
-
-    when:
-    String userMessage = exceptionMessageDecorator.decorate(runtimeException, runtimeException.getMessage())
-
-    then:
-    userMessage == runtimeException.getMessage()
-  }
-}
 
 class ExceptionMessageProvider implements ObjectProvider<List<ExceptionMessage>> {
 
@@ -104,8 +75,11 @@ class AccessDeniedExceptionMessage implements ExceptionMessage {
   }
 }
 
+@ResponseStatus(value = HttpStatus.FORBIDDEN, reason = "Access is denied")
 class LocalAccessDeniedException extends SpinnakerException {
+  LocalAccessDeniedException() {}
   LocalAccessDeniedException(String message) {
     super(message)
   }
 }
+
